+++ SharedPackages/BrowserServicesKit/Sources/Subscription/Managers/V1toV2Bridging/EntitlementsBridging.swift
@@ -22,22 +22,7 @@ import Networking
 public struct EntitlementsBridging {
 
     public static func v2EntitlementsFrom(v1Entitlements: [Entitlement]) -> [SubscriptionEntitlement] {
-        v1Entitlements.map { v1Entitlement in
-            switch v1Entitlement.product {
-            case .networkProtection:
-                return .networkProtection
-            case .dataBrokerProtection:
-                return .dataBrokerProtection
-            case .identityTheftRestoration:
-                return .identityTheftRestoration
-            case .identityTheftRestorationGlobal:
-                return .identityTheftRestorationGlobal
-            case .paidAIChat:
-                return .paidAIChat
-            case .unknown:
-                return .unknown
-            }
-        }
+        v1Entitlements.map { $0.product.subscriptionEntitlement }
     }
 }
 
@@ -77,6 +62,23 @@ extension SubscriptionEntitlement {
             return .paidAIChat
         case .unknown:
             return .unknown
+        }
+    }
+
+    var entitlement: Entitlement {
+        switch self {
+        case .networkProtection:
+            return Entitlement(product: .networkProtection)
+        case .dataBrokerProtection:
+            return Entitlement(product: .dataBrokerProtection)
+        case .identityTheftRestoration:
+            return Entitlement(product: .identityTheftRestoration)
+        case .identityTheftRestorationGlobal:
+            return Entitlement(product: .identityTheftRestorationGlobal)
+        case .paidAIChat:
+            return Entitlement(product: .paidAIChat)
+        case .unknown:
+            return Entitlement(product: .unknown)
         }
     }
 }
